move app into components directory
